Improve on the DivingLog importer a bit
This adds a few fields that we parse, but more importantly it also tries
to dynamically decide if the sample temperatures and pressures are in
imperial or metric units.

Dirk suspects that DivingLog generally always does everything in metric,
and the reason why he has crazy sample data in imperial units (both for
pressure and temperature) may be due to a bug in the early Uemis
importer for DivingLog.

Which would actually make a lot more sense than DivingLog really being
so insane on purpose.

Anyway, Dirk's brother Jurgen seems to have everything in metric units,
which would be much saner.  Maybe we should throw away the support for
insane DivingLog files entirely, since it is possible that the only use
ever of the possible source of that bug was Dirk's use of the Uemis
importer.

But for now, we end up just guessing. Current guesses:

 - water temperature is below 32 dgC, so 32+ degrees is in Fahrenheit.

 - tank pressures are below 400 bar, so higher values than that must be
   psi.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>